fix: dealing with IS-2, renaming 2 packages from "userCRUDservice" to "user_service"
