--- src/main/java/com/userCRUDservice/dtos/CreateUserDto.java
@@ -1,4 +1,0 @@
-package com.userCRUDservice.dtos;
-
-public class CreateUserDto {
-}
