+++ src/main/java/com/user_service/DemoApplication.java
@@ -1,8 +1,7 @@
-package com.userCRUDservice;
+package com.user_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.jdbc.autoconfigure.DataSourceAutoConfiguration;
 
 @SpringBootApplication
 public class DemoApplication {
