+++ src/main/java/com/user_service/dtos/RefreshTokenRequestDto.java
@@ -1,4 +1,4 @@
-package com.userCRUDservice.dtos;
+package com.user_service.dtos;
 
 public class RefreshTokenRequestDto {
 }
