+++ src/main/java/com/user_service/handlers/UserAlreadyExists.java
@@ -1,4 +1,4 @@
-package com.userCRUDservice.handlers;
+package com.user_service.handlers;
 
 public class UserAlreadyExists extends RuntimeException {
     public UserAlreadyExists(String message){
