+++ src/main/java/com/user_service/repository/UserRepository.java
@@ -1,11 +1,10 @@
-package com.userCRUDservice.repository;
+package com.user_service.repository;
 
-import com.userCRUDservice.model.User;
+import com.user_service.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
-import java.util.UUID;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
