+++ src/main/java/com/user_service/service/mapper/Mapper.java
@@ -1,4 +1,4 @@
-package com.userCRUDservice.service.mapper;
+package com.user_service.service.mapper;
 
 public interface Mapper <F, T> {
     T map(F from);
