+++ src/test/java/com/user_service/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.userCRUDservice;
+package com.user_service;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
